refactor(api): source route filter types from the service layer (#726)
Service-backed routes (bookings, add-ons, insurance-options, fee-schedules,
locations, notifications) imported their *Filters request shapes straight from
repositories/types, leaking data-layer contracts into the controller layer and
contradicting the AGENTS.md "routes never import repositories" rule.

Add services/filters.ts — a documented service-layer re-export of those filter
contracts — and point the six routes at it, keeping the import direction
routes -> services -> repositories intact. Pure type-only move, no runtime change.

The DI-repo carve-out routes (regions, stats, vehicles) keep importing their
*Repository interface from repositories/types as a constructor contract until
their own service extraction lands.

Part of #726.
